+++ flutter_inappwebview/example/integration_test/in_app_webview/main.dart
@@ -56,6 +56,7 @@ part 'load_data.dart';
 part 'load_file.dart';
 part 'load_file_url.dart';
 part 'load_file_url_request.dart';
+part 'save_state.dart';
 part 'load_url.dart';
 part 'on_console_message.dart';
 part 'on_content_size_changed.dart';
@@ -220,6 +221,7 @@ void main() {
     navigationListener();
     visitedHistory();
     conversationContext();
+    saveStateBounds();
     keepAlive();
 
     // `printCurrentPage` MUST BE LAST, and must stay last.